Fix comparison for negative kv in font formatting

+++ src/font/fmt_txt/lv_font_fmt_txt.c
@@ -247,7 +247,7 @@ bool lv_font_get_glyph_dsc_fmt_txt(const lv_font_t * font, lv_font_glyph_dsc_t *
     uint32_t adv_w = gdsc->adv_w;
     if(is_tab) adv_w *= 2;
 
-    if((kv < 0) && (adv_w < -kv)) {
+    if((kv < 0) && (adv_w < (uint32_t)-kv)) {
         adv_w = 0;
     }
     else {
